Fix bugs in serializers and tests
There was a bug in RegisterVerifyTests.test_success that was including
the prefix "Bearer " in the query string, which resulted in a false
positive. The prefix was removed from the query string in both
`test_success` and `test_invalid_token`.

To account for the difference between query strings and Authorization
headers, an overridden `validate_access` was added to
RegisterVerifySerializer. In both instances of `validate_access`,
another bug was corrected by raising ValidationError where and error
message was simply listed when it should have been included in an
exception.