Refactor check_urls/add mixins/update login tests
The primary purpose of this commit is to refactor the check_urls method
into a solution that doesn't involve comparing the same two strings
multiple times. To achieve this, a separate test has been created, which
takes two strings defined in the setUp method. The first string is the
hard-coded path, and the second is the name to be passed into the
django.urls.reverse function. LoginTests is the first to implement these
changes, and any other tests are assumed to be dysfunctional (which will
be fixed in the next commits).

Rather than defining this new method in LoginTests, it has been added to
a mixin. Since TestCaseShortcutsMixin was quickly growing, it is split
into subsidiary mixins, although the newly empty TestCaseShortcutsMixin
inherits from all of them.

Since the previous commit made `check()` functions within tests
obsolete, they have been integrated into their parent methods. Now,
nothing is passed to `self.check_urls()` since it no longer exists and
therefore does not need to be called.

The 'user_gone' check in LoginSerializer has been removed because
execution will never reach scope of `if not authenticated_user:` since
the get user method is guaranteed to populate a value if
User.DoesNotExist is not raised, and that exception is already being
handled.