Add dict app and TimestampedModel; update User
The dictionary app has been initiated. This is where the interface with
the Merriam-Webster API and the icons database will reside.

The API root now has a models.py file including TimestampedModel, which
is an abstract model including timestamps for when a model is created
and was last updated. It is now a parent of User, and User no longer
explicitly defines these two fields.

Since `is_superuser` is included in PermissionsMixin, it has been
removed from the User class declaration. A migration has been generated
as a result.

Each app's apps.py has been updated to use BigAutoField and to inherit
from api.