Organize Word models and utility classes
The original purpose of this commit was to extend ExternalAPIManager to
include MP3 downloads, but it made more sense to split calls to
different APIs (or media servers, in the case of MP3s).
ExternalAPIManager has been converted to a base class for some new
classes: DictionaryAPIManager, ThesaurusAPIManager, and
DictionaryAudioManager. There are changes to static variables,
decorators, and names of methods originally in ExternalAPIManager, but
the core functionality has remained the same.

Now, instead of a single file for Word, WordEntry, DictionaryEntry, and
ThesaurusEntry definitions, a module has been created including Word and
the newly created WordManager in one file and the rest in another. This
is to keep everything modular and therefore easy to locate. The tradeoff
is increased chance of cyclic dependencies, which can usually be
resolved by moving imports into the methods where they are needed.